+++ src/constants/selectOptions.js
@@ -25,6 +25,7 @@ export const OPTIONS = {
   CHECK_OUT_NOTE: 'CHECK_OUT_NOTE',
   ELECTRONIC_ACCESS_URI: 'ELECTRONIC_ACCESS_URI',
   ELECTRONIC_ACCESS_URL_RELATIONSHIP: 'ELECTRONIC_ACCESS_URL_RELATIONSHIP',
+  ELECTRONIC_ACCESS_LINK_TEXT: 'ELECTRONIC_ACCESS_LINK_TEXT',
 };
 
 export const PARAMETERS_KEYS = {
@@ -201,6 +202,12 @@ export const getHoldingsOptions = (formatMessage, holdingsNotes = []) => [
   {
     value: OPTIONS.ELECTRONIC_ACCESS_URL_RELATIONSHIP,
     label: formatMessage({ id: 'ui-bulk-edit.layer.options.holdings.urlRelationship' }),
+    disabled: false,
+    categoryName: formatMessage({ id: 'ui-bulk-edit.category.electronicAccess' }),
+  },
+  {
+    value: OPTIONS.ELECTRONIC_ACCESS_LINK_TEXT,
+    label: formatMessage({ id: 'ui-bulk-edit.layer.options.holdings.linkText' }),
     disabled: false,
     categoryName: formatMessage({ id: 'ui-bulk-edit.category.electronicAccess' }),
   },